gve: add a completion mechanism for mailbox messages
When sending messages to the devices via the mailbox, use the
completion API to wait for responses before timing out. A 16
bit 'cookie' field in the descriptor is used to associate
responses for sent mailbox messages. The driver fills this
out when sending a message.

This cookie field is mirrored back by the device in the response
message so that it can be used as an index to find the exact mbx
message that the driver got the response for and that message can
be 'completed'.

The size of the message queue is set to be smaller than the length of
the mailbox queue size to always have space for asynchronous messages
from the device.

Signed-off-by: Harshitha Ramamurthy <hramamurthy@google.com>