gve: implement GET_INTERRUPT_DBS mailbox message
Doorbell information is kept slightly differently for AQ mode vs mailbox
mode. In AQ mode, the reference is kept as a pointer to the index of the
doorbell within BAR2 (all registers in BAR2 are assumed to be 4 bytes in
width), while in mailbox mode, it is instead a direct offset into BAR0.

Mailbox mode also makes use of two separate registers: the IRQ doorbell
register and the IRQ coalescing register. In AQ mode, the functionality
of these registers is all kept in the single IRQ doorbell register.

irq_db_index is renamed to aq_irq_db_index to make it clear that this
register is AQ-specific. This message exposes the mapping between MSI-X
vectors and their corresponding IRQ and coalescing doorbells.

A batching-based approach is used to request interrupt doorbells. In the
request, a starting MSI-X index and the number of vectors for which to
request IRQ register offsets. The number requested is bounded from above
by the amount of doorbells that would fit in a response message. If more
register offsets are needed than would fit within a single response
buffer, the message can be called repeatedly with different starting
offsets.

In the RX processing step, register offsets are recorded in
pre-allocated notify block structs for later use when interrupts are
enabled. Resolve the IRQ doorbell MMIO address during notify block
creation and cache it directly in the gve_notify_block structure as
`irq_db`. This abstracts the control plane mode away from doorbell
accesses.

Signed-off-by: Joshua Washington <joshwash@google.com>
Signed-off-by: Harshitha Ramamurthy <hramamurthy@google.com>